Remove old ClientTuple
